Modify float to double

+++ camera.h
@@ -13,12 +13,12 @@ vec3 random_in_unit_disk() {
 
 class camera {
 	public:
-		camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov,
-				float aspect, float aperture, float focus_dist) {
+		camera(vec3 lookfrom, vec3 lookat, vec3 vup, double vfov,
+				double aspect, double aperture, double focus_dist) {
 			lens_radius = aperture / 2;
-			float theta = vfov * M_PI / 180; // thetaをラジアンに変える
-			float half_height = tan(theta / 2);
-			float half_width = aspect * half_height;
+			double theta = vfov * M_PI / 180; // thetaをラジアンに変える
+			double half_height = tan(theta / 2);
+			double half_width = aspect * half_height;
 			origin = lookfrom;
 			w = unit_vector(lookfrom - lookat);
 			u = unit_vector(cross(vup, w));
@@ -29,7 +29,7 @@ class camera {
 			horizontal = 2 * half_width * focus_dist * u;
 			vertical = 2 * half_height * focus_dist * v;
 		}
-		ray get_ray(float s, float t) {
+		ray get_ray(double s, double t) {
 			vec3 rd = lens_radius * random_in_unit_disk();
 			vec3 offset = u * rd.x() + v * rd.y();
 			return ray(origin + offset, lower_left_corner + s * horizontal + t * vertical - origin - offset);
@@ -40,7 +40,7 @@ class camera {
 		vec3 horizontal;
 		vec3 vertical;
 		vec3 u, v, w;
-		float lens_radius;
+		double lens_radius;
 };
 
 #endif
