+++ ray.h
@@ -9,7 +9,7 @@ class ray {
 			: orig(origin), dir(direction) {}
 		vec3 origin() const { return orig; }
 		vec3 direction() const { return dir; }
-		vec3 at(float t) const { return orig + t * dir; }
+		vec3 at(double t) const { return orig + t * dir; }
 
 		point3 orig;
 		vec3 dir;
